+++ src/views/Payments.vue
@@ -18,8 +18,8 @@
                             <table>
                                 <thead>
                                 <tr>
-                                    <th>{{$t('select')}}</th>
-                                    <th>ID</th>
+                                    <th><i class="fa fa-check"></i></th>
+                                    <th>{{$t('id')}}</th>
                                     <th>{{$t('amount')}}</th>
                                     <th>{{$t('created_at')}}</th>
                                 </tr>
@@ -87,6 +87,9 @@
             },
             loading () {
                 return this.$store.state.loading
+            },
+            language() {
+                return this.$store.state.language
             }
         },
         data () {
@@ -108,6 +111,9 @@
                 } else {
                     this.page = old; //FIXME find another way of blocking update when loading
                 }
+            },
+            language(val, old) {
+                moment.locale(val)
             }
         },
         filters: {
@@ -137,12 +143,12 @@
             en: {
                 amount: "Amount",
                 created_at: "Created",
-                select: "Select"
+                id: "ID"
             },
             pl: {
                 amount: "Kwota",
                 created_at: "Utworzono",
-                select: "Zaznaczenie"
+                id: "ID"
             }
         }
     }